flask_api/requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-Login==0.6.2
Flask-SocketIO==5.3.4
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.28.1
psycopg2-binary==2.9.6
eventlet==0.33.3
cryptography==41.0.2
requests==2.31.0
python-dotenv==1.0.0
pika==1.3.2